better-backend/requirements.txt
pillow
flask
rembg
openai
